backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Task Queue
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Browser Automation
playwright==1.41.2

# OCR and Document Processing
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
PyPDF2==3.0.1

# PDF Generation
reportlab==4.1.0

# AI Providers
openai==1.12.0
anthropic==0.18.1
httpx==0.26.0

# Object Storage
boto3==1.34.34
minio==7.2.3

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
orjson==3.9.13

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
